requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
google-generativeai>=0.4.0
gspread>=6.0.0
pillow>=10.0.0
python-dateutil>=2.8.0
pytest>=7.0.0

